staging-installcheck: move the InstallChecker class to its own loadable module
